Moved detectors and added naive unit tests in new dir
